+++ package.nix
@@ -11,7 +11,7 @@
 
 rustPlatform.buildRustPackage (finalAttrs: {
   pname = "dfdisk";
-  version = "0.1.4";
+  version = "0.1.5";
 
   src = lib.fileset.toSource {
     root = ./.;