+++ packaging/nixpkgs/package.nix
@@ -13,7 +13,7 @@
 
 rustPlatform.buildRustPackage (finalAttrs: {
   pname = "dfdisk";
-  version = "0.1.4";
+  version = "0.1.5";
 
   __structuredAttrs = true;